Conflicts:
	KendoDP2/Areas/Evaluacion360/Models/DP2Context.cs
	KendoDP2/Areas/Organizacion/Models/DP2Context.cs

+++ KendoDP2/Areas/Reclutamiento/Models/EvaluacionXFaseXPostulacion.cs
@@ -21,7 +21,7 @@ namespace KendoDP2.Areas.Reclutamiento.Models
         public DateTime FechaInicio { get; set; }
         public DateTime FechaFin { get; set; }
         
-        public int Puntaje { get; set; }
+        public double Puntaje { get; set; }
         public bool FlagAprobado { get; set; }
         
         public string Comentarios { get; set; }
@@ -60,7 +60,7 @@ namespace KendoDP2.Areas.Reclutamiento.Models
 
         public string FechaInicio { get; set; }
         public string FechaFin { get; set; }
-        public int Puntaje { get; set; }
+        public double Puntaje { get; set; }
         public bool FlagAprobado { get; set; }
         public string Comentarios { get; set; }
         public string Observaciones { get; set; }
